Convert RP to interface, add implementation
